Add streamelements command scraping
We can now set up an automatic commands scraping from streamelements.

Said commands (only :reply commands) will be exposed through a route
specified in the config file, grouped by common "entity" that comes
after prefix and before a number.

For example, if a prefix specified to be "insult" and there is a bunch
of commands such as ["insultcat" "insultcat1" "insultcat2"],
["insultorphan1" "insultorphan2"], etc., a route will
be exposed (with a specified :endpoint) that will return a random :reply
command. Route requires a `table' parameter which specifies the entity.

In addition, some streamelements macros are supported in a limited form,
such as ${random.pick}, ${user}, ${touser}, ${sender}. The latter three
require additional parameters supplied to route, "0" corresponds to
"user" and "touser", while "1" will be what will be substituted for "sender".